# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "lumino 记账本服务",
        "title": "Lumino",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/transaction": {
            "get": {
                "tags": [
                    "交易记录"
                ],
                "summary": "获取交易记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "对应的账本id",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "起始时间",
                        "name": "begin_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "交易记录",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "交易记录"
                ],
                "summary": "修改交易记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "对应的账本id",
                        "name": "account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "交易数额,数值需要\u003e0",
                        "name": "amount",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "关联消费场景分类ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "记录账户id",
                        "name": "creator_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "交易日期",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "注释",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "交易记录id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "RelatedUserIDs 涉及那些人",
                        "name": "related_user_ids",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "类型:收入/支出,交易类型必须是0或者1",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "交易记录"
                ],
                "summary": "注册交易记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "对应的账本id",
                        "name": "account_book_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "交易数额",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "关联消费场景分类ID",
                        "name": "category_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "创建人id",
                        "name": "creator_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "日期",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "注释",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "RelatedUserIDs 涉及那些人",
                        "name": "related_user_ids",
                        "in": "query"
                    },
                    {
                        "enum": [
                            1,
                            2
                        ],
                        "type": "integer",
                        "description": "类型:收入/支出",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "交易记录"
                ],
                "summary": "删除交易记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "账本id",
                        "name": "account_book_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "交易记录id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Transaction": {
            "type": "object",
            "properties": {
                "account_book_id": {
                    "description": "对应的账本id",
                    "type": "integer"
                },
                "amount": {
                    "description": "交易数额",
                    "type": "number"
                },
                "category_id": {
                    "description": "关联消费场景分类ID",
                    "type": "integer"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator_id": {
                    "description": "创建人id",
                    "type": "integer"
                },
                "date": {
                    "description": "日期",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "DeletedAt 删除时间",
                    "type": "string",
                    "example": "2025-03-26T00:00:00Z"
                },
                "description": {
                    "description": "注释",
                    "type": "string"
                },
                "id": {
                    "description": "主键id",
                    "type": "integer"
                },
                "related_user_ids": {
                    "description": "RelatedUserIDs 涉及那些人",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "type": {
                    "description": "类型:收入/支出",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}